+++ packages/boot-vue/rollup.config.js
@@ -1,0 +1,23 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import esmImportToUrl from 'rollup-plugin-esm-import-to-url';
+
+export default [
+  {
+    input: 'index.js',
+    treeshake: true,
+    plugins: [esmImportToUrl({
+      imports: {
+        '@alt-javascript/cdi': 'https://cdn.jsdelivr.net/npm/@alt-javascript/cdi@3/dist/alt-javascript-cdi-esm.js',
+      },
+    })],
+    output: {
+      file: 'dist/alt-javascript-boot-vue-esm.js',
+      format: 'esm',
+      strict: false,
+      externalLiveBindings: false,
+      freeze: false,
+      sourcemap: false,
+      sourcemapExcludeSources: true,
+    },
+  },
+];